chore(typedoc): refactor CLI cmds into separate modules
