+++ packages/docutils/lib/cli/command/index.ts
@@ -1,0 +1,3 @@
+export {default as init} from './init';
+export {default as validate} from './validate';
+export {default as build} from './build';
